+++ client/src/containers/Navbar/Navbar_Info.js
@@ -1,4 +1,3 @@
-// PAGE URLs
 const HOME_PATH = "/";
 const ABOUT_PATH = "/about";
 const CAREERS_PATH = "/careers";
@@ -10,16 +9,17 @@ const BROTHERS_PATH = "/brothers";
 const EXECUTIVES_PATH = "/eboard";
 const CONTACT_PATH = "/contact";
 
-const NAVBAR_MAP = {};
-NAVBAR_MAP[ABOUT_PATH] = ABOUT_PATH;
-NAVBAR_MAP[CAREERS_PATH] = ABOUT_PATH;
-NAVBAR_MAP[RECRUITMENT_PATH] = RECRUITMENT_PATH;
-NAVBAR_MAP[WHYPSE_PATH] = RECRUITMENT_PATH;
-NAVBAR_MAP[FAQS_PATH] = RECRUITMENT_PATH;
-NAVBAR_MAP[APPLICATION_PATH] = RECRUITMENT_PATH;
-NAVBAR_MAP[BROTHERS_PATH] = BROTHERS_PATH;
-NAVBAR_MAP[EXECUTIVES_PATH] = BROTHERS_PATH;
-NAVBAR_MAP[CONTACT_PATH] = CONTACT_PATH;
+const NAVBAR_MAP = {
+  ABOUT_PATH,
+  CAREERS_PATH,
+  RECRUITMENT_PATH,
+  WHYPSE_PATH,
+  FAQS_PATH,
+  APPLICATION_PATH,
+  BROTHERS_PATH,
+  EXECUTIVES_PATH,
+  CONTACT_PATH
+};
 
 const NAVBAR_INFO = [
   {
